libgimpcolor/gimprgb-parse.c use a signed integer to pass the string
2004-07-22  Sven Neumann  <sven@gimp.org>

	* libgimpcolor/gimprgb-parse.c
	* libgimpcolor/gimprgb.h: use a signed integer to pass the string
	length to the new parser functions. The API explicitely asks for
	-1 to be passed...

	* app/core/gimp.c
	* app/core/gimpgradient-load.[ch]
	* app/core/gimpgradient.h: added preliminary support for loading
	simple SVG gradients (see bug #148127).  Be careful with this new
	feature; editing the loaded gradient will cause the SVG file to be
	overwritten! Work in progress...

+++ libgimpcolor/gimprgb.h
@@ -53,13 +53,13 @@ void      gimp_rgb_get_uchar   (const GimpRGB *rgb,
 
 gboolean  gimp_rgb_parse_name  (GimpRGB       *rgb,
                                 const gchar   *name,
-                                gsize          len);
+                                gint           len);
 gboolean  gimp_rgb_parse_hex   (GimpRGB       *rgb,
                                 const gchar   *hex,
-                                gsize          len);
+                                gint           len);
 gboolean  gimp_rgb_parse_css   (GimpRGB       *rgb,
                                 const gchar   *css,
-                                gsize          len);
+                                gint           len);
 
 void      gimp_rgb_add         (GimpRGB       *rgb1,
                                 const GimpRGB *rgb2);
